Move files around to keep things more organized
