+++ db/controllers/mongo/seed_large_mongo_database.js
