+++ server/routers/restaurants_api.js
@@ -4,7 +4,7 @@ var getRestaurantById = require('../../db/controllers/getRestaurantById.js');
 
 router.get('/:id/sidebar', (req, res) => {
   var restaurantId = req.params.id;
-  getRestaurantById(restaurantId).then((result) => {
+  getRestaurantById.searchMongo(restaurantId).then((result) => {
     res.send(result);
   });
 });
